# Optimize recursive calls